Updated redux to use slices
