+++ client/src/App.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import {ReduxTestComponent} from './components/ReduxTestComponent'
 
 function App() {
   return (
@@ -18,7 +17,7 @@ function App() {
           rel="noopener noreferrer"
         >
           Learn React
-        </a>
+        </a>:
       </header>
     </div>
   );
