+++ client/src/components/ReduxTestComponent.tsx
@@ -1,10 +1,9 @@
 
 import { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { setData2 } from "../redux/TestFeature/reducer";
+import { setData2 } from "../redux/TestFeature/TestFeatureSlice";
 
 const ReduxTestComponent = () => {
-  const [count, setCount] = useState(0);
   const temp = useSelector((state: any ) => state.testFeature.data2);
 
   const dispatch = useDispatch();
@@ -14,7 +13,6 @@ const ReduxTestComponent = () => {
   }
 
   return (
-    
     <div>
         <button onClick={()=>incrementValue(temp+1)}>test</button>
         {temp}
